+++ db/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from db.hash import Hash
 from db.models import DbUser
 from schemas import UserBase, UserBaseForPatch
+from fastapi import HTTPException, status
 
 
 def create_user(db: Session, request: UserBase):
@@ -21,13 +22,16 @@ def get_all_users(db: Session):
 
 
 def get_user_by_id(db: Session, id: int):
-    # Handle any exceptions
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with {id} not found!")
+    return user
 
 
 def update_user_with_changes(db: Session, id: int, changes: dict):
-    # Handle any exceptions
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with {id} not found!")
     user.update(changes)
     db.commit()
     return user.first()
@@ -52,8 +56,9 @@ def update_user_password(db: Session, id: int, password: str):
 
 
 def delete_user(db: Session, id: int):
-    # Handle any exceptions
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with {id} not found!")
     db.delete(user)
     db.commit()
     return "OK"
